analyzer: Move package manager tests to the "managers" package
To correspond to the code being tested.
