+++ analyzer/src/funTest/kotlin/managers/YarnTest.kt
@@ -17,11 +17,9 @@
  * License-Filename: LICENSE
  */
 
-package com.here.ort.analyzer
+package com.here.ort.analyzer.managers
 
-import com.here.ort.analyzer.managers.Npm
 import com.here.ort.downloader.VersionControlSystem
-import com.here.ort.model.config.AnalyzerConfiguration
 import com.here.ort.model.yamlMapper
 import com.here.ort.utils.normalizeVcsUrl
 import com.here.ort.utils.safeDeleteRecursively
@@ -37,8 +35,8 @@ import io.kotlintest.specs.WordSpec
 
 import java.io.File
 
-class NpmVersionUrlTest : WordSpec() {
-    private val projectDir = File("src/funTest/assets/projects/synthetic/npm-version-urls").absoluteFile
+class YarnTest : WordSpec() {
+    private val projectDir = File("src/funTest/assets/projects/synthetic/yarn").absoluteFile
     private val vcsDir = VersionControlSystem.forDirectory(projectDir)!!
     private val vcsUrl = vcsDir.getRemoteUrl()
     private val vcsRevision = vcsDir.getRevision()
@@ -46,23 +44,25 @@ class NpmVersionUrlTest : WordSpec() {
     override fun afterTest(testCase: TestCase, result: TestResult) {
         // Make sure the node_modules directory is always deleted from each subdirectory to prevent side-effects
         // from failing tests.
-        val nodeModulesDir = projectDir.resolve("node_modules")
-        val gitKeepFile = nodeModulesDir.resolve(".gitkeep")
-        if (nodeModulesDir.isDirectory && !gitKeepFile.isFile) {
-            nodeModulesDir.safeDeleteRecursively(force = true)
+        projectDir.listFiles().forEach {
+            if (it.isDirectory) {
+                val nodeModulesDir = File(it, "node_modules")
+                val gitKeepFile = File(nodeModulesDir, ".gitkeep")
+                if (nodeModulesDir.isDirectory && !gitKeepFile.isFile) {
+                    nodeModulesDir.safeDeleteRecursively(force = true)
+                }
+            }
         }
     }
 
     init {
-        "NPM" should {
-            "resolve dependencies with URLs as versions correctly" {
+        "yarn" should {
+            "resolve dependencies correctly" {
                 val packageFile = File(projectDir, "package.json")
-
-                val config = AnalyzerConfiguration(ignoreToolVersions = false, allowDynamicVersions = true)
-                val result = createNPM(config).resolveDependencies(listOf(packageFile))[packageFile]
+                val result = createYarn().resolveDependencies(listOf(packageFile))[packageFile]
                 val vcsPath = vcsDir.getPathToRoot(projectDir)
                 val expectedResult = patchExpectedResult(
-                    File(projectDir.parentFile, "npm-version-urls-expected-output.yml"),
+                    File(projectDir.parentFile, "yarn-expected-output.yml"),
                     definitionFilePath = "$vcsPath/package.json",
                     url = normalizeVcsUrl(vcsUrl),
                     revision = vcsRevision,
@@ -74,6 +74,6 @@ class NpmVersionUrlTest : WordSpec() {
         }
     }
 
-    private fun createNPM(config: AnalyzerConfiguration = DEFAULT_ANALYZER_CONFIGURATION) =
-        Npm("NPM", USER_DIR, config, DEFAULT_REPOSITORY_CONFIGURATION)
+    private fun createYarn() =
+        Yarn("Yarn", USER_DIR, DEFAULT_ANALYZER_CONFIGURATION, DEFAULT_REPOSITORY_CONFIGURATION)
 }
